backend/requirements.txt
fastapi
uvicorn
sse-starlette
torch
librosa
transformers
soundfile
psutil
distro
accelerate
